fix(media): 完善媒体上游契约
补齐 Grok generations 视频路由、参考素材归一化与轮询恢复，完善 Adobe 媒体 SKU、公开别名和 API 文档配置。\n\n兼容上游浮点 Unix 时间戳并保持对外整数秒，补充香蕉/Grok/视频转存相关回归测试。\n\n文档：docs/media-request-chain-audit.md、docs/video-task-routing.md、relay/channel/task/README.md\n配合：adobe2api 统一异步任务时间戳输出。

+++ relay/channel/task/oaivideo/vendors/seedance/adaptor.go
@@ -288,7 +288,7 @@ func (a *TaskAdaptor) ConvertToOpenAIVideo(task *model.Task) ([]byte, error) {
 	if task.FinishTime > 0 {
 		openAIVideo.CompletedAt = task.FinishTime
 	} else if dResp.CompletedAt > 0 {
-		openAIVideo.CompletedAt = dResp.CompletedAt
+		openAIVideo.CompletedAt = int64(dResp.CompletedAt)
 	}
 	videoURL := oaivideo.ExtractVideoURL(dResp)
 	if videoURL == "" {
